refactor(ProfileControllerTest, dto): change module and file name
