+++ apps/backend/src/main/java/scrumpledpaper/agiler/kanban/dto/IssueKanbanConfigUpdateReqDto.java
@@ -2,7 +2,7 @@ package scrumpledpaper.agiler.kanban.dto;
 
 import jakarta.validation.constraints.NotNull;
 
-public record IssueKanbanConfigReqDto(
+public record IssueKanbanConfigUpdateReqDto(
 	@NotNull(message = "칸반 설정 ID는 필수입니다.")
 	Long kanbanConfigId
 ) {}
